[GrB] Refactor
