+++ solutions/GraphBLAS/utils.cpp
@@ -69,61 +69,3 @@ report(const BenchmarkParameters &parameters, int iteration, const std::string &
     }
 
 }
-
-/*
- * DEBUG FUNCTIONS
- */
-
-void WriteOutDebugMatrix(const char *title, GrB_Matrix result) {
-    printf("%s:\n", title);
-    GrB_Index rows, cols;
-    ok(GrB_Matrix_nrows(&rows, result));
-    ok(GrB_Matrix_ncols(&cols, result));
-    double element;
-
-    for (GrB_Index i = 0; i < rows; i++) {
-        for (GrB_Index j = 0; j < cols; j++) {
-            GrB_Info info = ok(GrB_Matrix_extractElement_FP64(&element, result, i, j), false);
-
-            if (info == GrB_SUCCESS) {
-                printf("%g ", element);
-            } else if (info == GrB_NO_VALUE) {
-                // It is up to the user to determine what 'no value'
-                // means.  It depends on the semiring used.
-                printf("- ");
-            } else {
-                printf("Error! %s\n", GrB_error());
-            }
-
-        }
-        printf("\n");
-    }
-}
-
-void WriteOutDebugVector(const char *title, GrB_Vector result) {
-    printf("%s:\n", title);
-    GrB_Index size;
-    ok(GrB_Vector_size(&size, result));
-    double element;
-
-    for (unsigned int i = 0; i < size; i++) {
-        GrB_Info info = ok(GrB_Vector_extractElement_FP64(&element, result, i), false);
-
-        if (info == GrB_SUCCESS) {
-            if (element == UINT64_MAX) {
-                printf("inf ");
-            } else {
-                printf("%g ", element);
-            }
-
-        } else if (info == GrB_NO_VALUE) {
-            // It is up to the user to determine what 'no value'
-            // means.  It depends on the semiring used.
-            printf("- ");
-        } else {
-            printf("Error! %s\n", GrB_error());
-        }
-
-    }
-    printf("\n");
-}
